Add FluentContext and move LangLocId to shared types
