+++ Linguini.Shared/Types/Bundle/FluentDateTime.cs
@@ -1,0 +1,18 @@
+﻿namespace Linguini.Shared.Types.Bundle
+{
+    /// <summary>
+    /// Represents a specific date and time value that can be formatted based on localization or other formatting options.
+    /// </summary>
+    public class FluentDateTime
+    {
+        
+    }
+
+    /// <summary>
+    /// Number options as defined in <see href="https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/DateTimeFormat">MDN Intl.DateTimeFormat()</see>
+    /// </summary>
+    public record FluentDateTimeOptions
+    {
+        
+    }
+}
